Changed how class types are registered (#14)
## Description
Changed how class types are registered in the ClassLevelMaps to be
coherent with how they are searched.

## Example
A Class type would be added by the method VisitCtClass(...) in
LatteClassFirstPass as "LinkedList",
but would be searched for by the method visitCtInvocation(...) in
LatteTypeChecker as "LinkedList<String>" (for example), leading to a
miss.
Was fixed by using .getTypeErasure() on visitCtInvocation(...),
simplifying to "LinkedList" in every case

## Related Issue
#6 

## Type of change
- [x] Bug fix
- [ ] New feature
- [ ] Documentation update
- [ ] Code refactoring

## How Has This Been Tested?
App.java was run and the previously known bug was found.
Adding other instances of the same class but with different Types did
not affect the correctness of the output.